committed

+++ Kilo_to_Miles.java
@@ -2,7 +2,7 @@ import java.util.*;
 public class Kilo_to_Miles{
    public static void main(String[] args){
          Scanner sc=new Scanner(System.in);
-		 double kilometer=sc.nextInt();
+		 double kilometer=sc.nextDouble();
 		 double miles=kilometer*0.621371;
 		 System.out.println(miles);
    }
